+++ src/saison/Models/Beer/BeerInfo.cs
@@ -54,7 +54,7 @@ namespace Saison.Models.Beer
         public Brewery Brewery { get; set; }
 
         [JsonPropertyName("auth_rating")]
-        public int AuthRating { get; set; }
+        public float AuthRating { get; set; }
 
         [JsonPropertyName("wish_list")]
         public bool WishList { get; set; }
