+++ src/saison/Models/Beer/MediaBeer.cs
@@ -53,7 +53,7 @@ namespace Saison.Models.Beer
         public bool HasHad { get; set; }
 
         [JsonPropertyName("auth_rating")]
-        public int AuthRating { get; set; }
+        public float AuthRating { get; set; }
 
         [JsonPropertyName("wish_list")]
         public bool WishList { get; set; }
