+++ src/saison/Models/Brewery/Beer.cs
@@ -50,7 +50,7 @@ namespace Saison.Models.Brewery
         public bool OnList { get; set; }
 
         [JsonPropertyName("auth_rating")]
-        public int AuthRating { get; set; }
+        public float AuthRating { get; set; }
 
         [JsonPropertyName("wish_list")]
         public bool WishList { get; set; }
