+++ src/saison/Models/User/Wishlist/Beer.cs
@@ -44,7 +44,7 @@ public class Beer
     public bool HasHad { get; set; }
 
     [JsonPropertyName("auth_rating")]
-    public int AuthRating { get; set; }
+    public float AuthRating { get; set; }
 
     [JsonPropertyName("wish_list")]
     public bool WishList { get; set; }
